+++ www/SQLitePlugin.js
@@ -638,6 +638,42 @@
         name: SelfTest.DBNAME,
         location: 'default'
       }, function(db) {
+        return db.executeSql('SELECT UPPER("Test") AS upperText', [], function(resutSet) {
+          if (!resutSet.rows) {
+            SelfTest.finishWithError(errorcb, 'Missing resutSet.rows');
+            return;
+          }
+          if (!resutSet.rows.length) {
+            SelfTest.finishWithError(errorcb, 'Missing resutSet.rows.length');
+            return;
+          }
+          if (resutSet.rows.length !== 1) {
+            SelfTest.finishWithError(errorcb, "Incorrect resutSet.rows.length value: " + resutSet.rows.length + " (expected: 1)");
+            return;
+          }
+          if (!resutSet.rows.item(0).upperText) {
+            SelfTest.finishWithError(errorcb, 'Missing resutSet.rows.item(0).upperText');
+            return;
+          }
+          if (resutSet.rows.item(0).upperText !== 'TEST') {
+            SelfTest.finishWithError(errorcb, "Incorrect resutSet.rows.item(0).upperText value: " + (resutSet.rows.item(0).data) + " (expected: 'TEST')");
+            return;
+          }
+          delete db.openDBs[SelfTest.DBNAME];
+          delete txLocks[SelfTest.DBNAME];
+          return SelfTest.start3(successcb, errorcb);
+        }, function(select_err) {
+          return SelfTest.finishWithError(errorcb, "SELECT error: " + select_err);
+        });
+      }, function(open_err) {
+        return SelfTest.finishWithError(errorcb, "Open database error: " + open_err);
+      });
+    },
+    start3: function(successcb, errorcb) {
+      return SQLiteFactory.openDatabase({
+        name: SelfTest.DBNAME,
+        location: 'default'
+      }, function(db) {
         return db.sqlBatch(['CREATE TABLE TestTable(id integer primary key autoincrement unique, data);', ['INSERT INTO TestTable (data) VALUES (?);', ['test-value']]], function() {
           var firstid;
           firstid = -1;